% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_map.R
\name{bivariate_map}
\alias{bivariate_map}
\title{bivariate_map}
\usage{
bivariate_map(w, x, y, dim, pal, style, size_axis, scale_leg, x_leg, y_leg)
}
\arguments{
\item{w}{Is the sf object with the variables to map.}

\item{x}{Is the variable x in the sf object to be mapped.}

\item{y}{Is the variable x in the sf object to be mapped.}

\item{dim}{The dimensions of the palette, either 2 for a two-by-two palette or 3 for a three-by-three palette}

\item{pal}{A palette name; one of "Brown", "DkBlue", "DkCyan", "DkViolet", or "GrPink". As biscale::bi_legend.}

\item{style}{A string identifying the style used to calculate breaks. Currently supported styles are "quantile", "equal", "fisher", and "jenks"}

\item{size_axis}{Size of axis labels.}

\item{scale_leg}{Scales the grob relative to the rectangle defined by x, y, width, height. A setting of scale = 1 indicates no scaling.}

\item{x_leg}{The x location of the legend in the plot.}

\item{y_leg}{The y location of the legend in the plot.}
}
\value{
a ggplot object.
}
\description{
generate a bivariate map
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
