% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_by_agem.R
\name{cases_by_agem}
\alias{cases_by_agem}
\title{cases_by_agem}
\usage{
cases_by_agem(path, disease)
}
\arguments{
\item{path}{is the directory where the txt of dengue file exist.}

\item{disease}{is a string for define the vector borne disease. The option are "Paludismo", "Dengue" & "covid.}
}
\value{
a sf object. The municipality with the dengue cases by year.
}
\description{
this function left joint sf AGEM and the cases by municipality.
}
\examples{
1+1
}
