% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geocoden.R
\name{data_geocoden}
\alias{data_geocoden}
\title{data_geocoden}
\usage{
data_geocoden(infile, data, sinave_new)
}
\arguments{
\item{infile}{is the name of file create with \link[denhotspots]{subset_den}.}

\item{data}{is a string for define the addreses or data. if data TRUE, is return the dataset.}

\item{sinave_new}{is logical value. If is TRUE the subset ir for 2016-2019 dengue dataset, else the subset ir for 2008-2015 dengue dataset.}
}
\value{
a vector.
}
\description{
this function creates an address vector and replaces incorrect text.
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
