% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoden.R
\name{geocoden}
\alias{geocoden}
\title{geocoden}
\usage{
geocoden(infile, address)
}
\arguments{
\item{infile}{is the name of the location that is being geocoded and is used to be saved in a file with extension rds.}

\item{address}{is vector addresses.}
}
\value{
a file with extension rds.
}
\description{
this function geocodes the addresses of the sinave database using Geocoding API.
}
\details{
The function use the \link[ggmap]{geocode} of ggmap.
}
\note{
The code for geocoding was taken from the personal page of Shane Lynn \url{https://www.shanelynn.ie/massive-geocoding-with-r-and-google-maps/} and adapted to work with the sinave \url{http://www.sinave.gob.mx/} database.
}
\seealso{
\link[ggmap]{geocode}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
