% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gihi.R
\name{gihi}
\alias{gihi}
\title{gihi}
\usage{
gihi(x, gi_hi, id, dis, time, alpha = NULL)
}
\arguments{
\item{x}{it is an sf object with the number of events for each spatial unit}

\item{gi_hi}{Define the local spatial statistic. It includes three options ("gi", "hi", "gi_hi", if "gi" calculates the spatial local statistic getis & ord, if it is "hi" calculates hi, and if it is "gi_hi" calculates both statistic}

\item{id}{is or are the identifiers of each spatial unit}

\item{dis}{it is the prefix of the variable name. Example den_2018 or denv_2018, dis is den and denv}

\item{time}{is the unit of time or the time scale of the analyzed time series. example in years, times, months, weeks}

\item{alpha}{is the value of alpha to determine the threshold of value of gi to discriminate hotspots from non-hotspots}
}
\value{
a sf object with the ids and the intensity and hotspots for hi, gi or each.
}
\description{
Calculate gi and hi local spatial statistic
}
\references{
Getis A, Ord JK. 1992. The analysis of spatial association by the use of distance statistics.Geographical Analysis, 24(2):189-206.
Ord JK, Getis A. 1995. Local spatial autocorrelation statistics: distributional issues and an application. Geographical Analysis, 27, 286–306.
Ord JK,  Getis A. 2012. Local spatial heteroscedasticity (LOSH), The Annals of Regional Science, 48 (2), 529–539.
Bivand RS, Wong DWS. 2018. Comparing implementations of global and local indicators of spatial association. TEST, 27(3), 716–748.
}
\seealso{
\link[spdep]{LOSH.cs}, \link[spdep]{localG}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
