% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_plot.R
\name{intensity_plot}
\alias{intensity_plot}
\title{intensuty_plot}
\usage{
intensity_plot(
  x,
  pal,
  option = NULL,
  name = NULL,
  pal_name,
  breaks,
  dir_pal = NULL,
  x_leg,
  y_leg
)
}
\arguments{
\item{x}{is the hotspots dataset.}

\item{pal}{is the palette.}

\item{option}{is the palette option.}

\item{name}{is the palette name. the option and name are very similar, some package use name (ej. rcartocolor::carto_pal) and other option (ex. viridis:viridis)}

\item{pal_name}{is a logical value, if pal_name TRUE is for packages that use name, else the packages that use option.}

\item{breaks}{is a numeric value. Is the increment of the sequence.}

\item{dir_pal}{is the direction of the palette, 1 forward, -1 reverse.}

\item{x_leg}{is the x coordinates of legend.}

\item{y_leg}{is the x coordinates of legend.}
}
\value{
a ggplot
}
\description{
this function plot the event hotspots by polygon.
}
\examples{
1+1 
}
