% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knox.R
\name{knox}
\alias{knox}
\title{knox test}
\usage{
knox(x, crs, dt, ds, sym, sp_link, planar_coord)
}
\arguments{
\item{x}{is the dataset with the coordinates and onset of symptoms}

\item{crs}{is the coordinate reference system.}

\item{dt}{is the temporal windows.}

\item{ds}{is the spatial windows.}

\item{sym}{is the MonteCarlo replications number.}

\item{sp_link}{is the logical value. If sp_link is true the space-time link is  built with sp, else with sf.}

\item{planar_coord}{is the logical value.if planar_coord is false, the dataset is projected to a planar system.}
}
\value{
a list with three elements: knox, st_link, space-time link.
     - *knox* is a list with three elements:
     \item{knox}{Knox statistic, which is the number of pairs of points found in a given space-time distance.}
     \item{p_value}{p-value calculated from MonteCarlo simulation.}
     \item{RR}{Relative Risk - calculated by observed value (Knox statistics) divided by mean of simulated values.}
     - *st_link* is a origen-destination dataset of class dataframe. This dataset has four variables:
     \item{Xo}{is the longitude of the point of origin.}
     \item{Yo}{are the latitud of the point of origin.}
     \item{Xd}{are the longitude of the destination point.}
     \item{Yd}{are the latitud of the destination point.}
     - *space-time link* is a sf object (LINESTRING) with crs 4326.
}
\description{
This function implement space-time analysis with the knox test
}
\details{
This function implements the space-time analysis with the knox test. x is a dataframe with three variables: x (longitude), y (latitude) and onset (onset of symptoms)
     The coordinates should be projected to a planar system.
     The function rknox test use the (knox)[https://github.com/thanhleviet/knox] package of (github)[https://github.com/].
}
\references{
\strong{Knox, E. (1964)}. \emph{The detection of space-time interactions.} Journal of the Royal Statistical Society. Series C (13(1), 25-30.
\strong{Tango, T. (2010)}. \emph{Statistical methods for disease clustering. Springer.}
}
\seealso{
\link[knox]{knox_mc}, \link[knox]{st_link}, \link[knox]{knox}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
