% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_risk.R
\name{map_risk}
\alias{map_risk}
\title{map_risk}
\usage{
map_risk(risk, staticmap)
}
\arguments{
\item{risk}{is the spatial risk dataset.}

\item{staticmap}{is a logical value, if true the map is static else the map is interactive.}
}
\value{
a ggplot o mapview object
}
\description{
The map_risk function generates the map of the operational scenarios or epidemiological scenarios based on the hotspots of dengue transmission and the hotspots of the abundance of dengue vector eggs.
}
\note{
To generate the map with this function, you must first run the risk_ageb function.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
