% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_to_polygons.R
\name{point_to_polygons}
\alias{point_to_polygons}
\title{point_to_polygon}
\usage{
point_to_polygons(x, y, ids, time, coords, crs, dis)
}
\arguments{
\item{x}{is the sf areal data.}

\item{y}{is the point pattern data with the coordinate.}

\item{ids}{is the unique o uniques identifiers of each spatial unit.}

\item{time}{is the temporal resolution of datasets.}

\item{coords}{is the coordinates of geocoded dataset.}

\item{crs}{is the coordinate reference system.}

\item{dis}{is the name of vector-borne diseases.}
}
\value{
a sf object.
}
\description{
this function counts how many events for spatial unit
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
