% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_law_plot.R
\name{power_law_plot}
\alias{power_law_plot}
\title{power_law_plot}
\usage{
power_law_plot(x)
}
\arguments{
\item{x}{is the dataset.}
}
\value{
a ggplot.
}
\description{
this plot generate the all Power Law for hotspots and spatial units.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
