% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power_laws}
\alias{power_laws}
\title{Power Laws data.}
\format{
A dataframe object with 606 rows and 3 variables:
\describe{
  \item{y}{are the percentiles of the y variable }
  \item{x}{are the percentiles of the x variable}
  \item{power_law}{are the power distributions}
  ...
}
}
\source{
\url{https://www.tandfonline.com/doi/abs/10.1080/00207390802213609}
}
\usage{
power_laws
}
\description{
A dataset containing the Power Laws dataset.
}
\keyword{datasets}
