% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_den_dataset.R
\name{read_den_dataset}
\alias{read_den_dataset}
\title{Read the dengue dataset for geocoding}
\usage{
read_den_dataset(x, des_edo_res, des_mpo_res)
}
\arguments{
\item{x}{is the path of dengue dataset.}

\item{des_edo_res}{is a string for define the state in uppercase.}

\item{des_mpo_res}{is a string for define the municipality in uppercase}
}
\value{
a data.table
}
\description{
Read the dengue dataset for geocoding
}
\examples{
1+1
}
