% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_h3.R
\name{risk_h3}
\alias{risk_h3}
\title{risk_h3}
\usage{
risk_h3(betas, hotspots, intensity_perc, locality, cve_edo)
}
\arguments{
\item{betas}{It is the dataset of the regression coefficients of the geostatistical model with INLA-SPDE. The betas are calculated with the deneggs package.}

\item{hotspots}{It is the database of the results of the hotspots analysis with the local statistician Getis&Ord. Hotspots are calculated with the denhotspots package.}

\item{intensity_perc}{It is the percentage of intensity of egg hotspots.}

\item{locality}{is the locality name.}

\item{cve_edo}{is the id of state.}
}
\value{
a sf object.
}
\description{
risk h3 identifies epidemiological scenarios based on historical epidemiological and entomological information.
}
\seealso{
\link[deneggs]{eggs_hotspots}, \link[deneggs]{spde_pred_map}, \link[deneggs]{eggs_hotspots_week} & @seealso \link[INLA]{inla}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
