% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscanR.R
\name{satscanR}
\alias{satscanR}
\title{satscanR}
\usage{
satscanR(
  cas_file,
  geo_file,
  start_date,
  end_date,
  spatial_window,
  temporal_window,
  ssbatchfilename = "SaTScanBatch",
  path_satscan
)
}
\arguments{
\item{cas_file}{is the case file (cas for short).}

\item{geo_file}{is the geographic or coordinates file (geo for short).}

\item{start_date}{is the earliest date to be included in the study period. Is calculated with min(x_cas$date).}

\item{end_date}{is the latest date to be included in the study period. Is calculated with max(x_cas$date).}

\item{spatial_window}{is the spatial window. The spatial resolution is in km.}

\item{temporal_window}{is the temporal window. The time resolution is in days.}

\item{ssbatchfilename}{is the name of the file containing the SaTScan executable.}

\item{path_satscan}{is the path of directory of SatScan installation.}
}
\value{
a list with eight object (main, col, rr, gis, llr, sci, shapeclust, prm)
}
\description{
This function performs space-time analysis with SaTScan.
}
\details{
the stascanR implements ths space-time analysis with [rsatscan](https://cran.r-project.org/web/packages/rsatscan/vignettes/rsatscan.html) package. The rsatscan package only does anything useful if you have installed [SaTScan](https://www.satscan.org/)
}
\seealso{
\link[rsatscan]{ss.options}, \link[rsatscan]{satscan}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}.
}
