% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_geocoden.R
\name{save_geocoden}
\alias{save_geocoden}
\title{save_geocoden}
\usage{
save_geocoden(x, y, directory, loc)
}
\arguments{
\item{x}{is the geocoded data.}

\item{y}{is the dataset o sinave.}

\item{directory}{is the folder where the RData file is to be saved}

\item{loc}{is the name of geocoded name locality}
}
\value{
a RData
}
\description{
This function generates the RData file where it contains the geocoded data and the sinave database.
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
