% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_joint_cov.R
\name{sf_joint_cov}
\alias{sf_joint_cov}
\title{sf_joint_cov
this function apply the dplyr::left_join between sf and covid data of México}
\usage{
sf_joint_cov(path_dat, path_sf)
}
\arguments{
\item{path_dat}{is the directory of the covid data.}

\item{path_sf}{is the directory of the municipality shapefile of Mexico.}
}
\value{
a sf object with a dengue cases accumulated by municipality.
}
\description{
sf_joint_cov
this function apply the dplyr::left_join between sf and covid data of México
}
\details{
The function use the \link[dplyr]{left_join} of dplyr.
}
\note{
the dataset of SARS-COV2 of Mexico is downloaded homepage of DGE \url{https://www.gob.mx/salud/documentos/datos-abiertos-152127?idiom=es} and adapted to work with the sinave \url{http://www.sinave.gob.mx/} database.
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
