% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticmap_intensity.R
\name{staticmap_intensity}
\alias{staticmap_intensity}
\title{staticmap_intensity}
\usage{
staticmap_intensity(
  x,
  pal,
  option = NULL,
  name = NULL,
  pal_name,
  breaks,
  dir_pal = NULL,
  x_leg,
  y_leg,
  ageb
)
}
\arguments{
\item{x}{is the hotspots dataset.}

\item{pal}{is the palette.}

\item{option}{is the palette option.}

\item{name}{is the palette name. the option and name are very similar, some package use name (ej. rcartocolor::carto_pal) and other option (ex. viridis:viridis)}

\item{pal_name}{is a logical value, if pal_name TRUE is for packages that use name, else the packages that use option.}

\item{breaks}{is a numeric value. Is the increment of the sequence.}

\item{dir_pal}{is the direction of the palette, 1 forward, -1 reverse.}

\item{x_leg}{is the x coordinates of legend.}

\item{y_leg}{is the x coordinates of legend.}

\item{ageb}{is a logical value, if ageb is TRUE plot the intensity map of ageb, else plot the intensity map by other spatial unit.}
}
\value{
a ggplot map.
}
\description{
This function generate the intensity map of hotspots.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}.
}
