% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticmap_satscan.R
\name{staticmap_satscan}
\alias{staticmap_satscan}
\title{staticmap_satscan}
\usage{
staticmap_satscan(
  x,
  rsatscan = NULL,
  satscan,
  locality,
  cve_edo,
  path_shapeclust = NULL,
  path_gis = NULL
)
}
\arguments{
\item{x}{is the dengue cases of the  target locality.}

\item{rsatscan}{is the output of the space-time analysis with [rsatscan](https://cran.r-project.org/web/packages/rsatscan/vignettes/rsatscan.html) package.}

\item{satscan}{is a logical value for indicating if TRUE for output of [SaTScan](https://www.satscan.org/).}

\item{locality}{is the target locality.}

\item{cve_edo}{is the id of state.}

\item{path_shapeclust}{is the directory of the output of the space-time analysis with [SaTScan](https://www.satscan.org/). is the col.shp file.}

\item{path_gis}{is the directory of the output of the space-time analysis with [SaTScan](https://www.satscan.org/).is the gis.shp file.}
}
\value{
a ggplot map.
}
\description{
This function generate the map of the output for [SaTScan](https://www.satscan.org/) or [rsatscan](https://cran.r-project.org/web/packages/rsatscan/vignettes/rsatscan.html).
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}.
}
