% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_den.R
\name{subset_den}
\alias{subset_den}
\title{subset_den}
\usage{
subset_den(x = NULL, path = NULL, edo, mun, loc, sinave_new, age_group, name)
}
\arguments{
\item{x}{is the 2008-2015 dengue dataset. Is NULL for 2016-2019 dengue dataset.}

\item{path}{is the path of the 2016-2019 dengue dataset. Is NULL for 2008-2015 dengue dataset.}

\item{edo}{is a string for define the state in uppercase.}

\item{mun}{is a string for define the municipality in uppercase.}

\item{loc}{is a string for define the locality in uppercase.}

\item{sinave_new}{is logical value. If is TRUE the subset ir for 2016-2019 dengue dataset, else the subset ir for 2008-2015 dengue dataset.}

\item{age_group}{is a logical value. If is TRUE the subset for age group (<= 12 & >= 65), else all age group.}

\item{name}{is the name the csv output}
}
\value{
a object names x and a csv file in working dircetory.
}
\description{
is a function for subset de dengue dataset
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
