% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_chain_map.R
\name{transmission_chains_map}
\alias{transmission_chains_map}
\title{transmission_chain_map}
\usage{
transmission_chains_map(geocoded_dataset, cve_edo, locality, dengue_cases)
}
\arguments{
\item{geocoded_dataset}{is the dengue geocoded dataset.}

\item{cve_edo}{is the id of state.}

\item{locality}{is the target locality}

\item{dengue_cases}{is string for define the positive of suspected dengue cases}
}
\value{
a mapview
}
\description{
the function generate the space-time links map with mapview package.
}
\examples{
1+1
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
