% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knox.R
\name{knox}
\alias{knox}
\title{knox test}
\usage{
knox(x, crs, dt, ds, sym, sp_link, planar_coord)
}
\arguments{
\item{x}{is the dataset with the coordinates and onset of symptoms}

\item{crs}{is the coordinate reference system.}

\item{dt}{is the temporal windows.}

\item{ds}{is the spatial windows.}

\item{sym}{is the MonteCarlo replications number.}

\item{sp_link}{is the logical value. If sp_link is true the space-time link is  built with sp, else with sf.}

\item{planar_coord}{is the logical value.if planar_coord is false, the dataset is projected to a planar system.}
}
\value{
a list with three elements: knox, st_link, space-time link.
     - *knox* is a list with three elements:
     \item{knox}{Knox statistic, which is the number of pairs of points found in a given space-time distance.}
     \item{p_value}{p-value calculated from MonteCarlo simulation.}
     \item{RR}{Relative Risk - calculated by observed value (Knox statistics) divided by mean of simulated values.}
     - *st_link* is a origen-destination dataset of class dataframe. This dataset has four variables:
     \item{Xo}{is the longitude of the point of origin.}
     \item{Yo}{are the latitud of the point of origin.}
     \item{Xd}{are the longitude of the destination point.}
     \item{Yd}{are the latitud of the destination point.}
     - *space-time link* is a sf object (LINESTRING) with crs 4326.
}
\description{
This function implement space-time analysis with the knox test
}
\details{
This function implements the space-time analysis with the knox test. x is a dataframe with three variables: x (longitude), y (latitude) and onset (onset of symptoms)
     The coordinates should be projected to a planar system.
     The function rknox test use the (knox)[https://github.com/thanhleviet/knox] package of (github)[https://github.com/].
}
\examples{
McHardy_dat <- data.frame(id = c(1:22),
x = c(266.8, 304.4, 265.5, 265.0, 264.2, 288.7, 290.2, 265.6, 263.7, 271.3, 267.4, 267.4, 276.5, 260.2, 264.0, 263.8, 300.5, 270.8, 258.7, 282.7,265.3, 285.3),
y = c(334.3, 379.3, 315.0, 314.0, 323.0, 265.2, 292.3, 318.2, 344.4, 333.5, 344.4, 344.4, 344.6, 358.2, 296.8, 344.3, 373.0, 326.1, 344.8, 322.3, 314.9, 261.0),
onset = c(1958, 1959, 1960, 1960, 1962, 1962, 1964, 1964, 1965, 1966, 1968, 1968, 1968, 1971, 1972, 1972, 1972, 1973, 1974, 1974, 1974, 1974))


rknox(x = McHardy_dat,
crs = 4979,
ds = 2,
dt = 0,
sym = 990,
sp_link  = TRUE,
planar_coord = TRUE)
}
\references{
\strong{Knox, E. (1964)}. \emph{The detection of space-time interactions.} Journal of the Royal Statistical Society. Series C (13(1), 25-30.
\strong{Tango, T. (2010)}. \emph{Statistical methods for disease clustering. Springer.}
}
\seealso{
\link[knox]{knox_mc}, \link[knox]{st_link}, \link[knox]{knox}
}
\author{
Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com}
}
